+++ docs/CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2024.10.22
+
+### Added
+
+- Date format checking.
+- SQL escaping.
+- Links in changelog.
+
+### Changed
+
+- DB operation function revamp.
+- Input handling revamp.
+- Improved Makefile.
+
 ### Fixed
 
 - get_input decrement bug.
@@ -66,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/OperaVaria/cdb-birthday/compare/1.3.0...HEAD
+[Unreleased]: https://github.com/OperaVaria/cdb-birthday/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/OperaVaria/cdb-birthday/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/OperaVaria/cdb-birthday/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/OperaVaria/cdb-birthday/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/OperaVaria/cdb-birthday/compare/1.0.0...1.1.0